Wave 3 working with significant reorg to use application view
