+++ src/app/views/contact_view.js
@@ -5,6 +5,8 @@ const ContactView = Backbone.View.extend({
   initialize: function(options){
     this.template = options.template;
     this.detailedTemplate = options.detailed;
+    // console.log(this.$el);
+    // this.on('click', this.closeIfOutOfFocus, this.$el.parent);
   },
   render: function(){
     this.delegateEvents();
@@ -21,10 +23,16 @@ const ContactView = Backbone.View.extend({
     this.render();
     console.log(this.model.toJSON());
     return this;
-  },
-  closeIfOutOfFocus: function(){
-
   }
+  // closeIfOutOfFocus: function(event){
+  //   console.log("WENT HERE");
+  //   var modalAndContents = $('#contact-details');
+  //
+  //   if (!modalAndContents.is(e.target) && modalAndContents.has(e.target).length == 0)
+  //   {
+  //     modalAndContents.hide();
+  //   }
+  // }
 });
 
 export default ContactView;
